+++ app/api/messages/direct-channel-send/route.ts
@@ -6,10 +6,10 @@ import { NextResponse } from "next/server";
 export async function POST(req: Request){
     try{
         const profile = await currentUserProfile();
-        const { content,checkMessageReplyId } = await req.json();
+        const { content,checkMessageReplyId,isReplyAPI } = await req.json();
         const { searchParams } = new URL(req.url);
     
-        const directChatIdProp = searchParams.get("directChatId");
+        const directChatIdProp = searchParams.get("channelId");
 
         if (!profile) {
           return new NextResponse("Unauthorized", { status: 401 });
@@ -49,6 +49,40 @@ export async function POST(req: Request){
             }
             }
         });
+
+        if(checkMessageReplyId !== "lmaoREPLY" && (checkMessageReplyId!== null || checkMessageReplyId !== undefined) && isReplyAPI === true)
+        {
+          const message2 = await db.directMessage.update({
+            where: {
+              id: message.id,
+              version: message.version
+            },
+            data: {
+              messageParentId: checkMessageReplyId,
+              version: {
+                increment: 1
+              },
+              isReply: true,
+              edited: false,
+              createdAt: new Date()
+            }
+          })
+          const message3 = await db.directMessage.findFirst({
+            where: {
+              id: message2.id,
+            },
+            include: {
+              userProfile: true,
+              messageParent: {
+                include: {
+                  userProfile: true
+                }
+              }
+            }
+          });
+          // console.log("testing reply", message3)
+         return NextResponse.json(message3);
+        }
 
         return NextResponse.json(message);
       } catch (error) {